moved minio fixture from conftest to test_integration so it does not get loaded in non-integration tests
